Migrate Cluster Controller to Kubebuilder v2
